[advance-override] Rename advance-proportional-override to advance-override
This is the follow up of crrev.com/c/2702291

Now that the CSSWG has standardized a descriptor 'advance-override' [1],
which works exactly the same as the 'advance-proportional-override', this
patch renames it to 'advance-override' to match the spec.

There's also a test moved from wpt_internal to wpt since it has been
standardized now.

[1] https://github.com/w3c/csswg-drafts/pull/5991

Bug: 1137633
Change-Id: I85d8cc5e3fef0315c36d1436cb21f1d4be85b04b